deduplicate timeout check
